# application/api-spec/src/main/resources/openapi/food-rescue.yaml
openapi: 3.0.3
info:
  title: Savr Platform API
  version: 0.0.1-SNAPSHOT
  description: |
    REST API for the Savr platform.
servers:
  - url: http://localhost:8080
...
paths:
...
  /offers:
    get:
      tags:
        - Offer
      summary: List offers
      description: |
        Returns available offers for browsing, or offers of a specific business when `businessId` is provided.

      operationId: getOffers
      parameters:
        - $ref: '#/components/parameters/BusinessIdQueryParam'
      responses:
        '200':
          description: Offers listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OfferResponse'
              examples:
                availableOffers:
                  value:
                    - id: 100
                      businessId: 10
                      title: Evening bakery rescue bag
                      description: Fresh pastries and bread from the end of the day
                      imageUrl: /images/offer-bakery.png
                      price: 4.90
                      quantityAvailable: 3
                      status: AVAILABLE
                      items:
                        - name: Bakery bundle
                          quantity: 1
                      pickupLocation:
                        address:
                          street: Hlavna 1
                          city: Kosice
                          postalCode: 04001
                          country: Slovakia
                        note: Back entrance near the parking lot
                      pickupTimeWindow:
                        from: 2025-04-08T18:00:00Z
                        to: 2025-04-08T19:30:00Z
                      createdAt: 2025-04-08T16:00:00Z
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access the requested business offers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    post:
      tags:
        - Offer
      summary: Create an offer
      description: |
        Creates and publishes a new offer for an active business owned by the authenticated user.

      operationId: createOffer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        description: Offer creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferRequest'
            examples:
              offerCreate:
                summary: New rescue offer
                value:
                  businessId: 10
                  title: Evening bakery rescue bag
                  description: Fresh pastries and bread from the end of the day
                  imageUrl: /images/offer-bakery.png
                  price: 4.90
                  quantityAvailable: 3
                  items:
                    - name: Bakery bundle
                      quantity: 1
                  pickupLocation:
                    address:
                      street: Hlavna 1
                      city: Kosice
                      postalCode: 04001
                      country: Slovakia
                    note: Back entrance near the parking lot
                  pickupTimeWindow:
                    from: 2025-04-08T18:00:00Z
                    to: 2025-04-08T19:30:00Z
      responses:
        '201':
          description: Offer created successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to create an offer for this business
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Business not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /offers/{id}:
    get:
      tags:
        - Offer
      summary: Get offer by id
      description: |
        Returns offer detail. Available offers can be viewed by authenticated users. Managed offers can also be viewed by their owner or admin.

      operationId: getOffer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OfferIdPathParam'
      responses:
        '200':
          description: Offer found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access this offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    put:
      tags:
        - Offer
      summary: Update offer
      description: |
        Updates an existing managed offer.

      operationId: updateOffer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OfferIdPathParam'
      requestBody:
        required: true
        description: Offer update payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfferRequest'
      responses:
        '200':
          description: Offer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to update this offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    delete:
      tags:
        - Offer
      summary: Cancel offer
      description: |
        Cancels an offer and removes it from active browsing.

      operationId: deleteOffer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OfferIdPathParam'
      responses:
        '204':
          description: Offer cancelled successfully
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to cancel this offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
  /notifications:
    get:
      tags:
        - Notification
      summary: List notifications
      description: |
        Returns the notification inbox of the authenticated user.

      operationId: getNotifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notifications listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /notifications/{id}:
    get:
      tags:
        - Notification
      summary: Get notification by id
      description: |
        Returns a notification of the authenticated user.

      operationId: getNotification
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationIdPathParam'
      responses:
        '200':
          description: Notification found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to access this notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /notifications/{id}/read:
    post:
      tags:
        - Notification
      summary: Mark notification as read
      description: |
        Marks a notification as read and returns the updated notification.

      operationId: markNotificationAsRead
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationIdPathParam'
      responses:
        '200':
          description: Notification marked as read successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          description: User is not allowed to modify this notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BusinessIdQueryParam:
      name: businessId
      in: query
      required: false
      description: Optional business identifier used to list offers of a managed business
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 10
...
    OfferIdPathParam:
      name: id
      in: path
      required: true
      description: Offer identifier
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 100
...
    NotificationIdPathParam:
      name: id
      in: path
      required: true
      description: Notification identifier
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 900
  responses:
    BadRequestResponse:
      description: Invalid request body or request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidBody:
              value:
                code: VALIDATION_ERROR
                message: Invalid request body
                details:
                  - name must not be blank
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
    UnauthorizedResponse:
      description: Authentication required or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: UNAUTHORIZED
                message: Full authentication is required to access this resource
                details: []
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
    InternalServerErrorResponse:
      description: Unexpected internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              value:
                code: INTERNAL_ERROR
                message: Unexpected internal error
                details:
                  - Please contact support
                timestamp: 2025-04-08T18:30:00Z
                path: /businesses
  schemas:
...
    CreateOfferRequest:
      type: object
      required:
        - businessId
        - title
        - price
        - quantityAvailable
        - items
        - pickupLocation
        - pickupTimeWindow
      properties:
        businessId:
          type: integer
          format: int64
          minimum: 1
          example: 10
        title:
          type: string
          minLength: 1
          maxLength: 160
          example: Evening bakery rescue bag
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Fresh pastries and bread from the end of the day
        imageUrl:
          type: string
          nullable: true
          maxLength: 1024
          example: /images/offer-bakery.png
        price:
          type: number
          format: double
          minimum: 0
          example: 4.90
        originalPrice:
          type: number
          format: double
          nullable: true
          minimum: 0
          example: 12.50
        quantityAvailable:
          type: integer
          minimum: 1
          example: 3
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OfferItem'
        pickupLocation:
          $ref: '#/components/schemas/PickupLocation'
        pickupTimeWindow:
          $ref: '#/components/schemas/PickupTimeWindow'
    UpdateOfferRequest:
      type: object
      required:
        - title
        - price
        - quantityAvailable
        - items
        - pickupLocation
        - pickupTimeWindow
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 160
          example: Updated bakery rescue bag
        description:
          type: string
          nullable: true
          maxLength: 1000
          example: Updated offer description
        imageUrl:
          type: string
          nullable: true
          maxLength: 1024
          example: https://cdn.example.com/offers/bakery-evening-bag.jpg
        price:
          type: number
          format: double
          minimum: 0
          example: 5.50
        originalPrice:
          type: number
          format: double
          nullable: true
          minimum: 0
          example: 13.90
        quantityAvailable:
          type: integer
          minimum: 1
          example: 2
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OfferItem'
        pickupLocation:
          $ref: '#/components/schemas/PickupLocation'
        pickupTimeWindow:
          $ref: '#/components/schemas/PickupTimeWindow'
    OfferResponse:
      type: object
      required:
        - id
        - businessId
        - title
        - price
        - quantityAvailable
        - status
        - items
        - pickupLocation
        - pickupTimeWindow
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 100
        businessId:
          type: integer
          format: int64
          example: 10
        title:
          type: string
          example: Evening bakery rescue bag
        description:
          type: string
          nullable: true
          example: Fresh pastries and bread from the end of the day
        imageUrl:
          type: string
          nullable: true
          example: /images/offer-bakery.png
        price:
          type: number
          format: double
          example: 4.90
        originalPrice:
          type: number
          format: double
          nullable: true
          example: 12.50
        quantityAvailable:
          type: integer
          example: 3
        status:
          $ref: '#/components/schemas/OfferStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OfferItem'
        pickupLocation:
          $ref: '#/components/schemas/PickupLocation'
        pickupTimeWindow:
          $ref: '#/components/schemas/PickupTimeWindow'
        createdAt:
          type: string
          format: date-time
          example: 2025-04-08T16:00:00Z
...
    NotificationResponse:
      type: object
      required:
        - id
        - userId
        - type
        - title
        - message
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 900
        userId:
          type: integer
          format: int64
          example: 2
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
          example: Reservation confirmed
        message:
          type: string
          example: Your reservation for "Evening bakery rescue bag" is active.
        createdAt:
          type: string
          format: date-time
          example: 2025-04-08T16:31:00Z
        readAt:
          type: string
          format: date-time
          nullable: true
          example: 2025-04-08T16:45:00Z
    OfferItem:
      type: object
      required:
        - name
        - quantity
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
          example: Bakery bundle
        quantity:
          type: integer
          minimum: 1
          example: 1
    PickupLocation:
      type: object
      required:
        - address
      properties:
        address:
          $ref: '#/components/schemas/Address'
        note:
          type: string
          nullable: true
          maxLength: 255
          example: Back entrance near the parking lot
    PickupTimeWindow:
      type: object
      required:
        - from
        - to
      properties:
        from:
          type: string
          format: date-time
          example: 2025-04-08T18:00:00Z
        to:
          type: string
          format: date-time
          example: 2025-04-08T19:30:00Z
    Address:
      type: object
      required:
        - street
        - city
        - postalCode
        - country
      properties:
        street:
          type: string
          minLength: 1
          maxLength: 255
          example: Hlavna 1
        city:
          type: string
          minLength: 1
          maxLength: 100
          example: Kosice
        postalCode:
          type: string
          minLength: 1
          maxLength: 20
          example: 04001
        country:
          type: string
          minLength: 1
          maxLength: 100
          example: Slovakia
        latitude:
          type: number
          format: double
          nullable: true
          minimum: -90
          maximum: 90
          example: 48.7205
        longitude:
          type: number
          format: double
          nullable: true
          minimum: -180
          maximum: 180
          example: 21.2578
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - details
        - timestamp
        - path
      properties:
        code:
          type: string
          example: CONFLICT
        message:
          type: string
          example: Business with the same name already exists for this owner
        details:
          type: array
          items:
            type: string
          example:
            - Duplicate or invalid database state
        timestamp:
          type: string
          format: date-time
          example: 2025-04-08T18:30:00Z
        path:
          type: string
          example: /businesses
...
    OfferStatus:
      type: string
      enum:
        - DRAFT
        - AVAILABLE
        - RESERVED
        - PICKED_UP
        - SOLD_OUT
        - EXPIRED
        - CANCELLED
...
    NotificationType:
      type: string
      enum:
        - BUSINESS_STATUS_CHANGED
        - OFFER_PUBLISHED
        - OFFER_RESERVED
        - OFFER_CANCELLED
        - RESERVATION_STATUS_CHANGED
        - SYSTEM